Remove settings navigation loading flashes
Problem:
The settings optimization suspended every tab behind a dynamic import. Opening Settings therefore displayed a visible Loading settings interstitial even though MCPs, Providers, Jobs, and Logs contain only small payloads.

Solution:
Load the lightweight settings tabs synchronously and isolate the deferred boundary around the CodeMirror editor alone. Render the Global Config shell immediately, preload its editor on hover or focus, and preserve editor geometry during an immediate module load without a textual interstitial.

Validation:
npm run typecheck
npm run lint
npm test (10 files, 35 tests)
npm audit --audit-level=low (0 vulnerabilities)
npm audit --omit=dev --audit-level=low (0 vulnerabilities)
npm run build
Bundle budget passed with 307 chunks; the only settings dynamic entry is GlobalConfigEditor.js at 331.29 kB.
git diff --check
