+++ src/pages/settings/Settings.lazy.test.tsx
@@ -34,7 +34,29 @@ vi.mock('./LogsTab', () => ({
     LogsTab: () => <div>Log settings loaded</div>,
 }));
 
-describe('lazy settings tabs', () => {
+describe('settings loading boundaries', () => {
+    it('renders lightweight tabs immediately without a loading interstitial', async () => {
+        const user = userEvent.setup();
+        render(
+            <MemoryRouter>
+                <Settings />
+            </MemoryRouter>,
+        );
+
+        expect(screen.getByText('MCP settings loaded')).toBeVisible();
+        expect(screen.queryByText(/Loading settings/)).not.toBeInTheDocument();
+
+        await user.click(screen.getByRole('button', { name: /Providers/ }));
+        expect(screen.getByText('Provider settings loaded')).toBeVisible();
+        expect(screen.queryByText(/Loading settings/)).not.toBeInTheDocument();
+
+        await user.click(screen.getByRole('button', { name: /Jobs/ }));
+        expect(screen.getByText('Job settings loaded')).toBeVisible();
+
+        await user.click(screen.getByRole('button', { name: /Logs/ }));
+        expect(screen.getByText('Log settings loaded')).toBeVisible();
+    });
+
     it('initializes CodeMirror only after Global Config is opened', async () => {
         const user = userEvent.setup();
         const { container } = render(
@@ -43,7 +65,7 @@ describe('lazy settings tabs', () => {
             </MemoryRouter>,
         );
 
-        expect(await screen.findByText('MCP settings loaded')).toBeVisible();
+        expect(screen.getByText('MCP settings loaded')).toBeVisible();
         expect(container.querySelector('.cm-editor')).not.toBeInTheDocument();
 
         await user.click(screen.getByRole('button', { name: /Global Config/ }));
